refactor: simplify CLI and daemon facades
