+++ src/daemon/artifact-download.ts
@@ -4,7 +4,6 @@ import https from 'node:https';
 import os from 'node:os';
 import path from 'node:path';
 import { AppError } from '../utils/errors.ts';
-import { runCmd } from '../utils/exec.ts';
 import { resolveTimeoutMs } from '../utils/timeouts.ts';
 
 const MAX_ARTIFACT_BYTES = 2 * 1024 * 1024 * 1024; // 2 GB
@@ -338,188 +337,5 @@ function decodeURIComponentSafe(value: string): string {
     return decodeURIComponent(value);
   } catch {
     return value;
-  }
-}
-
-// --- Archive extraction (merged from artifact-archive.ts) ---
-
-export async function extractTarInstallableArtifact(params: {
-  archivePath: string;
-  tempDir: string;
-  platform: 'ios' | 'android';
-  expectedRootName?: string;
-}): Promise<string> {
-  const rootName = await resolveTarArchiveRootName(params);
-  await runCmd('tar', ['xf', params.archivePath, '-C', params.tempDir]);
-  const installablePath = path.join(params.tempDir, rootName);
-  if (!fs.existsSync(installablePath)) {
-    throw new AppError(
-      'INVALID_ARGS',
-      `Expected extracted bundle "${rootName}" not found in archive`,
-    );
-  }
-  return installablePath;
-}
-
-export async function resolveTarArchiveRootName(params: {
-  archivePath: string;
-  platform: 'ios' | 'android';
-  expectedRootName?: string;
-}): Promise<string> {
-  const entriesResult = await runCmd('tar', ['-tf', params.archivePath], { allowFailure: true });
-  if (entriesResult.exitCode !== 0) {
-    throw new AppError('INVALID_ARGS', 'Artifact is not a valid tar archive', {
-      archivePath: params.archivePath,
-      stdout: entriesResult.stdout,
-      stderr: entriesResult.stderr,
-      exitCode: entriesResult.exitCode,
-    });
-  }
-
-  const entries = entriesResult.stdout
-    .split(/\r?\n/)
-    .map((entry) => entry.trim())
-    .filter(Boolean);
-  if (entries.length === 0) {
-    throw new AppError('INVALID_ARGS', 'Uploaded app bundle archive is empty');
-  }
-
-  const normalizedEntries = entries.map((entry) => normalizeArchiveEntry(entry));
-  const rootName =
-    params.expectedRootName ?? resolveArchiveRootName(normalizedEntries, params.platform);
-  const hasExpectedRoot = normalizedEntries.some(
-    (entry) => entry === rootName || entry.startsWith(`${rootName}/`),
-  );
-  if (!hasExpectedRoot) {
-    throw new AppError(
-      'INVALID_ARGS',
-      `Uploaded archive must contain a top-level "${rootName}" bundle`,
-    );
-  }
-
-  for (const entry of normalizedEntries) {
-    validateArchiveEntryPath(entry, rootName);
-  }
-
-  const verboseResult = await runCmd('tar', ['-tvf', params.archivePath]);
-  const lines = verboseResult.stdout.split(/\r?\n/).filter(Boolean);
-  for (const line of lines) {
-    const entryType = line[0];
-    if (entryType === 'l' || entryType === 'h') {
-      throw new AppError(
-        'INVALID_ARGS',
-        'Uploaded app bundle archive cannot contain symlinks or hard links',
-      );
-    }
-  }
-
-  return rootName;
-}
-
-export async function readZipEntries(archivePath: string): Promise<string[] | null> {
-  const result = await runCmd('unzip', ['-Z1', archivePath], { allowFailure: true });
-  if (result.exitCode !== 0) return null;
-  return result.stdout
-    .split(/\r?\n/)
-    .map((entry) => entry.trim())
-    .filter(Boolean);
-}
-
-function resolveArchiveRootName(entries: string[], platform: 'ios' | 'android'): string {
-  const roots = new Set<string>();
-  for (const entry of entries) {
-    const [root] = entry.split('/');
-    if (root) roots.add(root);
-  }
-  const rootEntries = [...roots];
-  if (platform === 'ios') {
-    const appRoots = rootEntries.filter((entry) => entry.toLowerCase().endsWith('.app'));
-    if (appRoots.length === 1) return appRoots[0];
-    if (appRoots.length === 0) {
-      throw new AppError(
-        'INVALID_ARGS',
-        'iOS app bundle archives must contain a single top-level .app directory',
-      );
-    }
-    throw new AppError(
-      'INVALID_ARGS',
-      `iOS app bundle archives must contain exactly one top-level .app directory, found: ${appRoots.join(', ')}`,
-    );
-  }
-  if (rootEntries.length === 1) return rootEntries[0];
-  throw new AppError(
-    'INVALID_ARGS',
-    `Archive must contain a single top-level bundle, found: ${rootEntries.join(', ')}`,
-  );
-}
-
-function normalizeArchiveEntry(entry: string): string {
-  if (entry.includes('\0')) {
-    throw new AppError('INVALID_ARGS', `Invalid archive entry: ${entry}`);
-  }
-  if (path.posix.isAbsolute(entry)) {
-    throw new AppError('INVALID_ARGS', `Archive entry must be relative: ${entry}`);
-  }
-  const normalized = path.posix.normalize(entry).replace(/^(\.\/)+/, '');
-  if (!normalized || normalized === '.' || normalized.startsWith('../')) {
-    throw new AppError('INVALID_ARGS', `Archive entry escapes bundle root: ${entry}`);
-  }
-  return normalized;
-}
-
-function validateArchiveEntryPath(entry: string, rootName: string): void {
-  if (entry !== rootName && !entry.startsWith(`${rootName}/`)) {
-    throw new AppError(
-      'INVALID_ARGS',
-      `Archive entry must stay inside top-level "${rootName}" bundle: ${entry}`,
-    );
-  }
-}
-
-// --- Upload handling (merged from upload.ts) ---
-
-export async function receiveUpload(
-  req: IncomingMessage,
-): Promise<{ artifactPath: string; tempDir: string }> {
-  const artifactType = req.headers['x-artifact-type'] as string | undefined;
-  const rawFilename = req.headers['x-artifact-filename'] as string | undefined;
-
-  if (!artifactType || !rawFilename) {
-    throw new AppError(
-      'INVALID_ARGS',
-      'Missing required headers: x-artifact-type and x-artifact-filename',
-    );
-  }
-  if (artifactType !== 'file' && artifactType !== 'app-bundle') {
-    throw new AppError(
-      'INVALID_ARGS',
-      `Invalid x-artifact-type: ${artifactType}. Must be "file" or "app-bundle".`,
-    );
-  }
-
-  validateArtifactContentLength(req.headers['content-length']);
-  const artifactFilename = sanitizeArtifactFilename(rawFilename);
-  const tempDir = createArtifactTempDir('upload');
-
-  try {
-    if (artifactType === 'file') {
-      const destPath = path.join(tempDir, artifactFilename);
-      await streamReadableToFile(req, destPath);
-      return { artifactPath: destPath, tempDir };
-    }
-
-    const archivePath = path.join(tempDir, 'artifact.tar');
-    await streamReadableToFile(req, archivePath);
-    const destPath = await extractTarInstallableArtifact({
-      archivePath,
-      tempDir,
-      platform: 'ios',
-      expectedRootName: artifactFilename,
-    });
-    fs.rmSync(archivePath, { force: true });
-    return { artifactPath: destPath, tempDir };
-  } catch (error) {
-    fs.rmSync(tempDir, { recursive: true, force: true });
-    throw error;
   }
 }
